requirements.txt
playwright.sync_api
re
time
configparser